backend/requirements.txt
# FastAPI and dependencies
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9
asyncpg==0.29.0

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2

# Task Queue
celery==5.3.4
redis==5.0.1

# Storage
boto3==1.34.25
minio==7.2.3

# Utilities
python-dotenv==1.0.0
aiofiles==23.2.1
httpx==0.26.0
slowapi==0.1.9
psutil==5.9.8

# Observability (optional but recommended for production)
sentry-sdk[fastapi,celery,sqlalchemy]==1.40.0
prometheus-fastapi-instrumentator==6.1.0

# AI providers (optional — install only the ones you use)
anthropic==0.18.1; python_version >= "3.8"
# openai==1.12.0  # uncomment to enable OpenAI provider

# Development
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
black==24.1.1
flake8==7.0.0
mypy==1.8.0
isort==5.13.2
autoflake==2.2.1
